Move "backtick identifiers" inner test class
